SSAFY_1학기/05_django/새 폴더/requirements.txt
asgiref==3.5.2
Django==3.2.18
djangorestframework==3.13.1
pytz==2022.2.1
sqlparse==0.4.2